AssemblyScript doesn't seem to like inheritance, so move the reward vault code into the Kodiak price handler
