+++ subgraphs/berachain/config.json
@@ -1,6 +1,6 @@
 {
-  "id": "QmXsfRqogVcSeboRjpZwHB9iEqsh5d1Nm1MxceVKc8nWDF",
+  "id": "QmWKhC2iJ8CD8PgeSX7k3Z9uCmwyD7btqq8drmw44YQKKn",
   "org": "olympusdao",
   "name": "protocol-metrics-berachain",
-  "version": "1.0.0"
+  "version": "1.0.3"
 }
